requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
djangorestframework-simplejwt
mysqlclient
pymysql
pillow
django-filter
drf-yasg
reportlab
openpyxl
pandas
